Contactly Backend/requirements.txt
blinker
certifi
charset-normalizer
click
colorama
et-xmlfile
Flask
greenlet
gunicorn
idna
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
numpy
openpyxl
packaging
pandas
Pillow
protobuf
Pygments
PyMySQL
python-dateutil
python-dotenv
pytz
requests
six
SQLAlchemy
typing_extensions
tzdata
urllib3
waitress
Werkzeug
